// projeto_logistica/src/types/types.ts
export type tSubcategoriaTransportadoras = {
    id: number;
    nome?: string | undefined;
    state: boolean;
}

export type tCategoriasDropdownTrasportadoras = {
    id: number;
    nome: string;
    state: boolean
}[]

export type tSubcategoriasDropdownTransportadoras = {
    id: number;
    nome: string;
}[]